Split up network command into seperate files
